# backend/src/routes/auth.ts
import { Router } from "express";
import { z } from "zod";
import { authenticateWithLdap } from "../auth/ldap.js";
import { signAccessToken, signRefreshToken, verifyRefreshToken } from "../auth/jwt.js";
import { authenticateLocalUser, upsertLdapUser } from "../db/users.js";
import { requireAuth } from "../middleware/requireAuth.js";
import { getDb } from "../db/mssql.js";
import sql from "mssql";

const LoginSchema = z
  .object({
    identifier: z.string().min(1).optional(),
    username: z.string().min(1).optional(),
    password: z.string().min(1),
    provider: z.enum(["ldap", "local"]).default("ldap"),
  })
  .transform((data) => ({
    identifier: (data.identifier ?? data.username ?? "").trim(),
    password: data.password,
    provider: data.provider,
  }));

export const authRouter = Router();

const loadUserRoles = async (userId: string): Promise<string[]> => {
  const db = await getDb();
  const result = await db
    .request()
    .input("userId", sql.UniqueIdentifier, userId)
    .query(
      [
        "SELECT r.Name AS RoleName",
        "FROM pm.UserRoles ur",
        "INNER JOIN pm.Roles r ON r.RoleId = ur.RoleId",
        "WHERE ur.UserId = @userId",
      ].join("\n"),
    );

  const rows = result.recordset as Array<{ RoleName?: unknown }>;
  return rows
    .map((row) => (typeof row.RoleName === "string" ? row.RoleName : null))
    .filter((value): value is string => Boolean(value));
};

authRouter.post("/login", async (req, res) => {
  const parsed = LoginSchema.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ message: "Invalid request" });
    return;
  }

  if (!parsed.data.identifier) {
    res.status(400).json({ message: "Invalid request" });
    return;
  }

  try {
    if (parsed.data.provider === "local") {
      const user = await authenticateLocalUser({
        identifier: parsed.data.identifier,
        password: parsed.data.password,
      });

      const accessToken = signAccessToken({
        sub: user.userId,
        username: user.username,
        roles: user.roles,
      });
      const refreshToken = signRefreshToken({
        sub: user.userId,
        username: user.username,
        roles: user.roles,
      });

      res.json({
        accessToken,
        refreshToken,
        user: {
          id: user.userId,
          username: user.username,
          displayName: user.displayName,
          email: user.email,
          roles: user.roles,
        },
      });
      return;
    }

    const profile = await authenticateWithLdap(parsed.data.identifier, parsed.data.password);
    const roles = profile.isSuperadmin ? ["Superadmin"] : ["Viewer"];
    const user = await upsertLdapUser({
      username: profile.username,
      displayName: profile.displayName,
      email: profile.email,
      phone: profile.phone,
      externalId: profile.dn,
      roles,
    });

    const accessToken = signAccessToken({
      sub: user.userId,
      username: user.username,
      roles: user.roles,
    });
    const refreshToken = signRefreshToken({
      sub: user.userId,
      username: user.username,
      roles: user.roles,
    });

    res.json({
      accessToken,
      refreshToken,
      user: {
        id: user.userId,
        username: user.username,
        displayName: user.displayName,
        email: user.email,
        roles: user.roles,
      },
    });
  } catch (err) {
    if (err && typeof err === "object" && "code" in err && err.code === "LDAP_NOT_CONFIGURED") {
      res.status(503).json({ message: "LDAP is not configured", code: "LDAP_NOT_CONFIGURED" });
      return;
    }
    res.status(401).json({ message: "Invalid username or password" });
  }
});

authRouter.get("/me", requireAuth, async (req, res) => {
  const db = await getDb();
  const result = await db
    .request()
    .input("userId", sql.UniqueIdentifier, req.user.sub)
    .query(
      [
        "SELECT TOP (1) Username, DisplayName, Email",
        "FROM pm.Users",
        "WHERE UserId = @userId",
      ].join("\n"),
    );

  const row = result.recordset[0] as { Username: unknown; DisplayName: unknown; Email: unknown } | undefined;
  const username = typeof row?.Username === "string" ? row.Username : req.user.username;
  const displayName = typeof row?.DisplayName === "string" ? row.DisplayName : null;
  const email = typeof row?.Email === "string" ? row.Email : null;

  let roles = req.user.roles;
  try {
    roles = await loadUserRoles(req.user.sub);
  } catch {
    roles = req.user.roles;
  }

  res.json({
    user: {
      id: req.user.sub,
      username,
      displayName,
      email,
      roles,
    },
  });
});

const RefreshSchema = z.object({ refreshToken: z.string().min(1) });

authRouter.post("/refresh", async (req, res) => {
  const parsed = RefreshSchema.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ message: "Invalid request" });
    return;
  }

  try {
    const claims = verifyRefreshToken(parsed.data.refreshToken);

    let roles = claims.roles;
    try {
      roles = await loadUserRoles(claims.sub);
    } catch {
      roles = claims.roles;
    }

    const accessToken = signAccessToken({ sub: claims.sub, username: claims.username, roles });
    const refreshed = signRefreshToken({ sub: claims.sub, username: claims.username, roles });
    res.json({ accessToken, refreshToken: refreshed });
  } catch {
    res.status(401).json({ message: "Unauthorized" });
  }
});

const ThemeModeSchema = z.enum(["dark", "light"]);
const PreferencesSchema = z.object({
  themeMode: ThemeModeSchema.nullable().optional(),
  themePalette: z.string().trim().max(64).nullable().optional(),
});

authRouter.get("/me/preferences", requireAuth, async (req, res) => {
  const db = await getDb();
  const result = await db
    .request()
    .input("userId", sql.UniqueIdentifier, req.user.sub)
    .query(
      [
        "SELECT ThemeMode, ThemePalette",
        "FROM pm.Users",
        "WHERE UserId = @userId",
      ].join("\n"),
    );

  const row = result.recordset[0] as { ThemeMode: unknown; ThemePalette: unknown } | undefined;
  const themeMode = typeof row?.ThemeMode === "string" ? row?.ThemeMode : null;
  const themePalette = typeof row?.ThemePalette === "string" ? row?.ThemePalette : null;
  res.json({ themeMode, themePalette });
});

authRouter.put("/me/preferences", requireAuth, async (req, res) => {
  const parsed = PreferencesSchema.safeParse(req.body);
  if (!parsed.success) {
    res.status(400).json({ message: "Invalid request" });
    return;
  }

  const db = await getDb();
  const themeMode = parsed.data.themeMode ?? null;
  const themePalette = parsed.data.themePalette ?? null;

  await db
    .request()
    .input("userId", sql.UniqueIdentifier, req.user.sub)
    .input("themeMode", sql.NVarChar(16), themeMode)
    .input("themePalette", sql.NVarChar(64), themePalette)
    .query(
      [
        "UPDATE pm.Users",
        "SET ThemeMode = @themeMode,",
        "    ThemePalette = @themePalette,",
        "    UpdatedAt = sysutcdatetime()",
        "WHERE UserId = @userId",
      ].join("\n"),
    );

  res.json({ ok: true });
});