# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: Preventive Pilot API
  version: 1.0.1
  description: REST API for Preventive Pilot (web + mobile clients). Docs updated 2026-09-27.
servers:
  - url: /
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserPreferences:
      type: object
      properties:
        themeMode:
          type:
            - string
            - "null"
        themePalette:
          type:
            - string
            - "null"
      required:
        - themeMode
        - themePalette
    UserPreferencesRequest:
      type: object
      properties:
        themeMode:
          type:
            - string
            - "null"
          enum:
            - dark
            - light
            - null
        themePalette:
          type:
            - string
            - "null"
          maxLength: 64
          description: Trimmed before storage.
...
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
      required:
        - ok
      additionalProperties: false
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
      additionalProperties: true
    LoginRequest:
      type: object
      properties:
        identifier:
          type: string
        username:
          type: string
        password:
          type: string
        provider:
          type: string
          enum:
            - ldap
            - local
          default: ldap
      required:
        - password
      additionalProperties: false
    LoginResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          type: object
          properties:
            id:
              type: string
            username:
              type: string
            displayName:
              type:
                - string
                - "null"
            email:
              type:
                - string
                - "null"
            roles:
              type: array
              items:
                type: string
          required:
            - id
            - username
            - roles
      required:
        - accessToken
        - refreshToken
        - user
    RefreshRequest:
      type: object
      properties:
        refreshToken:
          type: string
      required:
        - refreshToken
      additionalProperties: false
    RefreshResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
      required:
        - accessToken
        - refreshToken
      additionalProperties: false
    MeResponse:
      type: object
      properties:
        user:
          type: object
          properties:
            id:
              type: string
            username:
              type: string
            roles:
              type: array
              items:
                type: string
          required:
            - id
            - username
            - roles
      required:
        - user
...
security:
  - bearerAuth: []
...
paths:
...
  /api/auth/me/preferences:
    get:
      tags:
        - Auth
      summary: Get current user preferences
      description: Authenticated user only. Missing values or missing user row return null preferences.
      parameters: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPreferences"
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      tags:
        - Auth
      summary: Replace current user preferences
      description: >-
        Authenticated user only. Both preference columns are written: omitted fields become null; an empty
        object clears both preferences.
      parameters: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "400":
          description: Invalid request or state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserPreferencesRequest"
...
  /api/auth/login:
    post:
      tags:
        - Auth
      summary: Login and receive access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid username or password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: >-
            LDAP is not configured. Local login remains available; omitted login provider still defaults to
            LDAP.
          content:
            application/json:
              schema:
                type: object
                required:
                  - message
                  - code
                properties:
                  message:
                    type: string
                  code:
                    type: string
                    enum:
                      - LDAP_NOT_CONFIGURED
  /api/auth/refresh:
    post:
      tags:
        - Auth
      summary: Refresh access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/me:
    get:
      tags:
        - Auth
      summary: Get current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...